requirements.txt
customtkinter
tk
Pillow
pyperclip
pyinstaller
requests
nuitka
CTkToolTip
plyer; sys_platform == 'win32' or 'linux'
pyobjus; sys_platform == 'win32' or 'linux'
macos-notifications; sys_platform == 'darwin'